requirements.txt
contourpy==1.1.0
cycler==0.11.0
filelock==3.12.2
fonttools==4.40.0
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.25.0
packaging==23.1
Pillow==9.5.0
pyparsing==3.1.0
python-dateutil==2.8.2
six==1.16.0
sympy==1.12
torch==2.0.1
typing_extensions==4.6.3
